Add auto-configuration capabilities for TMC9660 motor and current sensing
- Introduced new `MotorProfile` and `AutoConfig` structures for streamlined motor and current sensing configuration.
- Implemented `configureAuto` methods for automatic setup of motor parameters, including type, direction, PWM frequency, and current sensing settings.
- Enhanced the `BLDCComprehensiveTest` to utilize the new auto-configuration features, improving test coverage and validation of motor setups.
- Updated documentation to reflect new configuration options and usage examples.

These changes significantly enhance the configurability and usability of the TMC9660 driver, facilitating easier integration and improved performance in motor applications.

+++ examples/esp32/main/TestFramework.h
@@ -180,16 +180,16 @@ struct TestResults {
    * @brief Calculate success percentage
    * @return Success percentage (0.0 to 100.0)
    */
-  double get_success_percentage() const noexcept {
-    return total_tests > 0 ? (static_cast<double>(passed_tests) / total_tests * 100.0) : 0.0;
+  float get_success_percentage() const noexcept {
+    return total_tests > 0 ? (static_cast<float>(passed_tests) / total_tests * 100.0f) : 0.0f;
   }
 
   /**
    * @brief Get total execution time in milliseconds
    * @return Total execution time in milliseconds
    */
-  double get_total_time_ms() const noexcept {
-    return total_execution_time_us / 1000.0;
+  float get_total_time_ms() const noexcept {
+    return total_execution_time_us / 1000.0f;
   }
 };
